+++ packages/expo-dev-launcher/android/src/debug/java/expo/modules/devlauncher/launcher/DevLauncherReactHost.kt
@@ -10,7 +10,6 @@ import com.facebook.react.defaults.DefaultComponentsRegistry
 import com.facebook.react.defaults.DefaultReactHostDelegate
 import com.facebook.react.defaults.DefaultTurboModuleManagerDelegate
 import com.facebook.react.fabric.ComponentFactory
-import com.facebook.react.interfaces.exceptionmanager.ReactJsExceptionHandler
 import com.facebook.react.runtime.JSCInstance
 import com.facebook.react.runtime.ReactHostImpl
 import com.facebook.react.runtime.hermes.HermesInstance
@@ -48,7 +47,6 @@ object DevLauncherReactHost {
         jsRuntimeFactory = jsRuntimeFactory,
         turboModuleManagerDelegateBuilder = DefaultTurboModuleManagerDelegate.Builder()
       )
-    val reactJsExceptionHandler = ReactJsExceptionHandler { _ -> }
     val componentFactory = ComponentFactory()
     DefaultComponentsRegistry.register(componentFactory)
     val useDeveloperSupport = launcherIp != null
@@ -57,7 +55,6 @@ object DevLauncherReactHost {
       defaultReactHostDelegate,
       componentFactory,
       useDeveloperSupport,
-      reactJsExceptionHandler,
       useDeveloperSupport
     )
       .apply {