+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/views/FilteredReadableMap.kt
@@ -41,8 +41,8 @@ class FilteredReadableMap(
   private val backingMap: ReadableMap,
   private val filteredKeys: List<String>
 ) : ReadableMap by backingMap {
-  override fun getEntryIterator(): Iterator<Map.Entry<String, Any>> =
-    FilteredIterator(backingMap.getEntryIterator()) {
+  override val entryIterator: Iterator<Map.Entry<String, Any>> =
+    FilteredIterator(backingMap.entryIterator) {
       !filteredKeys.contains(it.key)
     }