+++ packages/expo/CHANGELOG.md
@@ -27,6 +27,10 @@
 
 - Update react-native to 0.74.3. ([#30139](https://github.com/expo/expo/pull/30139) by [@gabrieldonadel](https://github.com/gabrieldonadel))
 
+### ⚠️ Notices
+
+- Added support for React Native 0.75.x. ([#30034](https://github.com/expo/expo/pull/30034) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+
 ## 51.0.15 - 2024-06-20
 
 _This version does not introduce any user-facing changes._